sched/core: Move newprio local to its usage
The __sched_setscheduler() uses a lot of local variables, some of those
however are required only in specific code blocks. When those variables are
initialized at the beginning of the function and used only further down,
that makes the code less readable. The newprio is one such variables, it's
used in just one code block which is 200 code lines down.

Keep code cleaner and easier to read by moving the newprio definition
closer to its usage.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>  
Cc: Peter Zijlstra <peterz@infradead.org>  
Cc: linux-kernel@vger.kernel.org