sched/core: move priority checks into dedicated __check_sched_params()
The __sched_setscheduler() method growth into a "monster" function which is very
difficult to parse from top to bottom while keeping track of all the
code blocks dependencies and ordering.

Start fixing that by moving some code blocks into dedicated and well scoped
functions. Do that by adding a new __check_sched_params() which is meant to
provide an "ordered" logic to verify if the request can be satisfied.

Move into the new function the checks that needs to be performed on all calls,
either from kernespace or userspace and independently from userspace permissions.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

---
The following patches will move other blocks in the new function. Let's to that
in steps to make it easier the review of the moved code.